% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_all.R
\name{CovCor_err}
\alias{CovCor_err}
\title{Title The correlation and covariance}
\usage{
CovCor_err(data, condition, plot.line = FALSE, plot.heat = FALSE)
}
\arguments{
\item{data}{Simulated data frame}

\item{condition}{A specific group level}

\item{plot.line}{Plot the line graph (TRUE or FALSE)}

\item{plot.heat}{Plot the heatmap (TRUE or FALSE)}
}
\description{
Title The correlation and covariance
}
