% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_all.R
\name{CovHeatPlot}
\alias{CovHeatPlot}
\title{Title Plot the variance-covariance matrix in heatmap}
\usage{
CovHeatPlot(CMatrix)
}
\arguments{
\item{CMatrix}{A squared matrix with numeric elements}
}
\description{
Title Plot the variance-covariance matrix in heatmap
}
