% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_all.R
\name{CovMarginal_sim}
\alias{CovMarginal_sim}
\title{Title Generate the mariginal variance matrix}
\usage{
CovMarginal_sim(CorrSub, SigmaGroup, CorrRep, SigmaRep, sigmaRE = NULL)
}
\arguments{
\item{CorrSub}{Subject level correlation matrix, dim nq*nq}

\item{SigmaGroup}{Group level variance, length q}

\item{CorrRep}{Correlation matrix between repeated measures, dim m*m}

\item{SigmaRep}{Variance of repeated measures, length m}
}
\value{
Marginal covariance matrix cross samples, dim nqm*nqm
}
\description{
Title Generate the mariginal variance matrix
}
