% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_all.R
\name{fVarAR1}
\alias{fVarAR1}
\title{Title AR(1) correlation matrix}
\usage{
fVarAR1(rho, d)
}
\arguments{
\item{rho}{Correlation coefficient}

\item{d}{Dimension of repeated measures}
}
\value{
AR(1) correlation matrix
}
\description{
Title AR(1) correlation matrix
}
