% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_all.R
\name{fVarCS}
\alias{fVarCS}
\title{Title CS correaltion matrix}
\usage{
fVarCS(rho, d)
}
\arguments{
\item{rho}{Correlation coefficient}

\item{d}{Dimension of repeated measures}
}
\value{
CS correlation matrix
}
\description{
Title CS correaltion matrix
}
