% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_all.R
\name{wrap_design}
\alias{wrap_design}
\title{Title
wrap_design <- function(n,m,q, SingleGroup = FALSE, eps = FALSE)}
\usage{
wrap_design(n, m, q, SingleGroup = FALSE, eps = FALSE)
}
\arguments{
\item{n}{Number of patients in each group. (Need >50 individuals to have a reasonable estimated covariance structure.)}

\item{m}{Number of repeated visits (time points).}

\item{q}{Number of groups (eg., control and multiple treatment groups).}

\item{SingleGroup}{Whether the individuals are in single group or multiple groups.}
}
\value{
Model matrix with desired design.
}
\description{
Title
wrap_design <- function(n,m,q, SingleGroup = FALSE, eps = FALSE)
}
