% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_all.R
\name{wrap_sim_corr}
\alias{wrap_sim_corr}
\title{Title Simulate DGE List}
\usage{
wrap_sim_corr(
  data,
  N,
  eps_Signal,
  SingleGroup,
  coef.beta0 = NULL,
  coef.beta1 = NULL,
  coef.beta2 = NULL,
  coef.int = NULL,
  sigma = NULL,
  mu = NULL,
  tau = NULL,
  CovMarginal = NULL,
  act.cor = NULL
)
}
\arguments{
\item{data}{Design matrix in long format.}

\item{N}{Number of simulated genes in total.}

\item{eps_Signal}{Proportion of signal genes.}

\item{coef.beta0}{Mean of the global effect.}

\item{coef.beta1}{Mean of group effect.}

\item{coef.beta2}{Mean of time effect.}

\item{coef.int}{Coefficient of interaction terms.}

\item{sigma}{Standard deviation of the simulated AR(1) sequence.}

\item{mu}{Mean vector of random intercept and slope.}

\item{tau}{Elements of variance-covariance matrix of random intercept and slope.}

\item{CovMarginal}{Marginal covariance.}

\item{act.cor}{Correlation coefficient of the time effect.}
}
\value{
DGE list
}
\description{
Title Simulate DGE List
}
