% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_all.R
\name{wrap_sim_DGEList}
\alias{wrap_sim_DGEList}
\title{Simulate Differential Gene Expression Data}
\usage{
wrap_sim_DGEList(
  n,
  m,
  q,
  N_gene,
  global_eff = NULL,
  trt_eff,
  trt_slope = NULL,
  inter = NULL,
  sigma = NULL,
  mu = NULL,
  tau = NULL,
  act.cor = NULL,
  distribution = NULL,
  SingleGroup = FALSE
)
}
\arguments{
\item{n}{Integer, the number of biological replicates per group.}

\item{m}{Integer, the number of time points.}

\item{q}{Integer, the number of groups.}

\item{N_gene}{Integer, the number of genes to simulate.}

\item{global_eff}{Numeric vector or NULL, global effects for each gene.}

\item{trt_eff}{Numeric vector, treatment effects for each gene.}

\item{trt_slope}{Numeric vector or NULL, treatment slope effects for each gene.}

\item{inter}{Numeric vector or NULL, interaction effects for each gene.}

\item{sigma}{Numeric vector or NULL, residual standard deviations for each gene.}

\item{mu}{Numeric vector or NULL, mean expression levels for each gene.}

\item{tau}{Numeric vector or NULL, overdispersion parameters for each gene (applicable if \code{distribution} is "NB").}

\item{act.cor}{Numeric vector or NULL, actual correlations for each gene.}

\item{distribution}{Character string or NULL, specifies the distribution to use for simulation. If "NB", a negative binomial distribution is used.}

\item{SingleGroup}{Logical, if TRUE, simulates data for a single group.}
}
\value{
A list with the following components:
\itemize{
\item \code{norm_counts}: Transposed normalized counts matrix with genes in rows and samples in columns.
\item \code{pheno_data}: Data frame containing phenotype data for the samples.
\item \code{feature_data}: Data frame containing feature (gene) data.
}
}
\description{
Simulates gene expression data for a given number of genes and samples, and returns a list
containing the normalized counts matrix, sample phenotype data, and feature (gene) data.
}
\note{
The function assumes that the necessary libraries (such as \code{dplyr}, \code{edgeR}, and any other required packages) are already loaded in the R session.
The actual simulation functions \code{wrap_simulation} and \code{wrap_simulation_NB} are not defined within this documentation and should be provided for the function to work.
}
\examples{
\dontrun{
simulated_DGEList <- wrap_sim_DGEList(n = 5, m = 3, q = 2, N_gene = 1000,
                                      global_eff = rep(0, 1000), trt_eff = rep(1, 1000),
                                      trt_slope = rep(0.5, 1000), inter = rep(0.1, 1000),
                                      sigma = rep(1, 1000), mu = rep(10, 1000),
                                      tau = rep(0.2, 1000), act.cor = rep(0.3, 1000),
                                      distribution = "NB", SingleGroup = FALSE)
}

}
\seealso{
\code{\link[edgeR]{DGEList}}
}
\author{
\link{Zhiwen Jiang and Tony Kam-Thong}
}
