% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_all.R
\name{wrap_simulation}
\alias{wrap_simulation}
\title{Title Simulate the log-scale expression of one gene}
\usage{
wrap_simulation(
  data,
  SingleGroup,
  beta0 = NULL,
  beta1 = NULL,
  beta2 = NULL,
  lambda = NULL,
  sigma = NULL,
  mu = NULL,
  tau = NULL,
  act.cor = NULL
)
}
\arguments{
\item{data}{Design matrix in long format.}

\item{beta0}{Mean of the global effect.}

\item{beta1}{Mean of group effect.}

\item{beta2}{Mean of time effect.}

\item{lambda}{Coefficient of interaction terms.}

\item{sigma}{Standard deviation of the simulated AR(1) sequence.}

\item{mu}{Mean vector of random intercept and slope.}

\item{tau}{Elements of variance-covariance matrix of random intercept and slope.}

\item{act.cor}{Correlation parameter of the AR(1) structure.}
}
\value{
Data matrix in long data frame.
}
\description{
Title Simulate the log-scale expression of one gene
}
